-Malzeme bazlı operasyonlar migration hataları giderildi
-Malzeme bazlı operasyonlar, şablon operasyonlara göre kod ve isim doldurma, satır doldurma işlemleri yapıldı

+++ TsiErp.Entities/Entities/ProductsOperationLine/Dtos/UpdateProductsOperationLinesDto.cs
@@ -32,7 +32,7 @@ namespace TsiErp.Entities.Entities.ProductsOperationLine.Dtos
         /// <summary>
         /// Ayar ve Kontrol Süresi
         /// </summary>
-        public int AdjustmentAndControlTime { get; set; }
+        public decimal AdjustmentAndControlTime { get; set; }
         /// <summary>
         /// Satır Numarası
         /// </summary>
